# Major refactor of deserialization to support concurrent readers for performance