--- a/tensorizer/stream_io.py
+++ b/tensorizer/stream_io.py
@@ -186,7 +186,7 @@ class CAInfo:
         return hash(self._curl_flags)
 
 
-class CURLStreamFile(io.RawIOBase):
+class CURLStreamFile(io.BufferedIOBase):
     """
     CURLStreamFile implements a file-like object around an HTTP download, the
     intention being to not buffer more than we have to. It is intended for
@@ -550,7 +550,7 @@ else:
     _MAX_TCP_BUFFER_SIZE = 16 << 20  # 16 MiB
 
 
-class RedisStreamFile:
+class RedisStreamFile(io.BufferedIOBase):
     """
     RedisStreamFile implements a file-like object around a Redis key namespace. Each
     'file' is broken up into multiple keys, each of which is a slice of the file. Each
@@ -588,6 +588,8 @@ class RedisStreamFile:
         self,
         uri: str,
         *,
+        begin: Optional[int] = None,
+        end: Optional[int] = None,
         buffer_size: Optional[int] = None,
     ) -> None:
         if buffer_size is None:
@@ -625,7 +627,6 @@ class RedisStreamFile:
         self._curr_buffer = bytearray(largest)
         self._curr_buffer_idx = -1
         self._curr_buffer_view = memoryview(self._curr_buffer)[0:0]
-        self.closed = False
 
         init_end = time.monotonic()
 
@@ -634,6 +635,9 @@ class RedisStreamFile:
         self.bytes_read = 0
         self.bytes_skipped = 0
         self.read_operations = 0
+
+        if begin:
+            self.seek(begin)
 
     def _find_key_index(self, position):
         for i, index in enumerate(self._indexes):
@@ -799,13 +803,13 @@ class RedisStreamFile:
         self._curr = position
 
     def close(self):
-        self.closed = True
         if self._redis is not None:
             self._redis.close()
             self._redis = None
         if self._redis_tcp is not None:
             self._redis_tcp.close()
             self._redis_tcp = None
+        super(RedisStreamFile, self).close()  # will set self.closed to True
 
     def readline(self):
         raise io.UnsupportedOperation("readline")
@@ -942,6 +946,8 @@ def s3_download(
     s3_endpoint: str = default_s3_read_endpoint,
     buffer_size: Optional[int] = None,
     force_http: bool = False,
+    begin: Optional[int] = None,
+    end: Optional[int] = None,
     certificate_handling: Optional[CAInfo] = None,
 ) -> CURLStreamFile:
     url = _s3_download_url(
@@ -956,6 +962,8 @@ def s3_download(
         url,
         buffer_size=buffer_size,
         certificate_handling=certificate_handling,
+        begin=begin,
+        end=end
     )
 
 
@@ -1096,6 +1104,8 @@ def open_stream(
     s3_config_path: Optional[Union[str, bytes, os.PathLike]] = None,
     buffer_size: Optional[int] = None,
     force_http: bool = False,
+    begin: Optional[int] = None,
+    end: Optional[int] = None,
     *,
     certificate_handling: Optional[CAInfo] = None,
 ) -> Union[CURLStreamFile, RedisStreamFile, typing.BinaryIO]:
@@ -1139,6 +1149,11 @@ def open_stream(
         force_http: If True, force the use of HTTP instead of HTTPS for
             S3 downloads. This will double the throughput, but at the cost
             of security.
+        begin: if specified, the file or request will begin at this byte offset.
+            Uses seek() for files and Range for HTTP and S3. This has no effect
+            on writes.
+        end: if specified, HTTP and S3 requests will use this as the end of the
+            Range header. This has no effect on files or on writes.
         certificate_handling: Customize handling of SSL CA certificates for
             HTTPS and S3 downloads.
             Pass None to use default certificate verification, or an instance of
@@ -1210,6 +1225,8 @@ def open_stream(
         return CURLStreamFile(
             path_uri,
             buffer_size=buffer_size,
+            begin=begin,
+            end=end,
             certificate_handling=certificate_handling,
         )
     elif scheme == "redis":
@@ -1217,7 +1234,12 @@ def open_stream(
             raise ValueError(
                 'Only the mode "rb" is valid when opening redis:// streams.'
             )
-        return RedisStreamFile(path_uri, buffer_size=buffer_size)
+        return RedisStreamFile(
+            path_uri,
+            buffer_size=buffer_size,
+            begin=begin,
+            end=end
+        )
 
     elif scheme == "s3":
         if normalized_mode not in ("br", "bw", "ab", "+bw", "+ab"):
@@ -1322,6 +1344,8 @@ def open_stream(
                 s3_endpoint,
                 buffer_size=buffer_size,
                 force_http=force_http,
+                begin=begin,
+                end=end,
                 certificate_handling=certificate_handling,
             )
             if error_context:
@@ -1334,11 +1358,12 @@ def open_stream(
                 'Only binary modes ("rb", "wb", "wb+", etc.)'
                 " are valid when opening local file streams."
             )
-        dirname = os.path.dirname(path_uri)
-        if dirname:
-            os.makedirs(os.path.dirname(path_uri), exist_ok=True)
+        if "w" in normalized_mode:
+            dirname = os.path.dirname(path_uri)
+            if dirname:
+                os.makedirs(os.path.dirname(path_uri), exist_ok=True)
         if buffer_size is None:
             buffer_size = io.DEFAULT_BUFFER_SIZE
         handle: typing.BinaryIO = open(path_uri, mode, buffering=buffer_size)
-        handle.seek(0)
+        handle.seek(begin or 0)
         return handle