… README.md
… examples/benchmark_buffer_size/benchmark.py
… examples/deserialize-simple.py
… examples/deserialize.py
… pyproject.toml
… requirements.txt
… tests/test_serialization.py