This gives the opportunity to deserialize much faster in cases where you can support multiple streams, either against local disk or via HTTP.

## Notable changes:
* 'plaid mode' does not exist. The behavior of plaid mode to reuse buffers in CUDA is now the only possible path when using CUDA. As long as we are synchronizing, this _is_ safe, efficient, and best way to transfer tensors to a GPU.
* `num_readers` is a new argument, defaulting to '1'. Major performance benefits can be realized when increasing num_readers to, say, 8. In this situation the file_obj must be a URI or path so that more readers can be opened.
* When num_readers > 1, read_tensors() and read_numpy_arrays() will yield items in any order, not necessarily the order they are requested
* file reads and cuda copies are perfomed on the same thread, x the number of readers. This is called the "copy thread"

## Backwards compatible?
I attempted to minimize the number of backwards-incompatible interface changes to maintain version compatibility. I think existing behavior is the same as long as `num_readers=1`, but open to alternate takes. And whether it's worth maintaining that compatibility or cutting a new major version.

## Results

Reading a 20B fp16 file off a local NVME: (file in kernel cache)

* `main`: 7.96s, 5.2GB/s
* this branch with 1 reader: 7.78s, 5.3GB/s
* this branch with 8 readers: 3.22s, 12.8GB/s
* this branch with 8 readers and `PYTORCH_CUDA_ALLOC_CONF=pinned_use_cuda_host_register:True`: 2.23s, 18.5 GB/s


```
$ python deserialize.py --source /scratch/gpt-neox-20b.tensors --model-ref EleutherAI/gpt-neox-20b --num-readers=8
Deserialized 41.3 GB in 3.22s, 12.8 GB/s
Memory usage before: CPU: (maxrss: 1,294MiB F: 968,185MiB) GPU: (U: 433MiB F: 45,162MiB T: 45,596MiB) TORCH: (R: 0MiB/0MiB, A: 0MiB/0MiB)
Memory usage after: CPU: (maxrss: 6,424MiB F: 963,042MiB) GPU: (U: 39,919MiB F: 5,676MiB T: 45,596MiB) TORCH: (R: 39,456MiB/39,456MiB, A: 39,382MiB/39,382MiB)
Output: ¡Hola! Encantado de conocerte. hoy voy a pasarte la info a que me das un precio. Deberia verificar los datos mas el lunes. pero queria saber si a ud. le intereza tener una reunión de dos horas sobre mis productos
to CUDA stats: 41293685880 bytes in 10.256s, 3.750 GiB/s
readinto stats: 41293685880 bytes in 7.226s, 5.322 GiB/s
```


```
$ PYTORCH_CUDA_ALLOC_CONF=pinned_use_cuda_host_register:True python deserialize.py --source /scratch/gpt-neox-20b.tensors --model-ref EleutherAI/gpt-neox-20b --num-readers=8
Deserialized 41.3 GB in 2.23s, 18.5 GB/s
Memory usage before: CPU: (maxrss: 1,435MiB F: 968,043MiB) GPU: (U: 433MiB F: 45,162MiB T: 45,596MiB) TORCH: (R: 0MiB/0MiB, A: 0MiB/0MiB)
Memory usage after: CPU: (maxrss: 6,564MiB F: 962,914MiB) GPU: (U: 39,919MiB F: 5,676MiB T: 45,596MiB) TORCH: (R: 39,456MiB/39,456MiB, A: 39,382MiB/39,382MiB)
Output: ¡Hola! Encantado de conocerte. hoy voy a estar ocupadísimo porque llega mi marido.

A la manera de un niño mimado la señora se pasó la cena dando al niño las cosas que más le gustaban, que
to CUDA stats: 41293685880 bytes in 7.099s, 5.417 GiB/s
readinto stats: 41293685880 bytes in 6.679s, 5.758 GiB/s
```