@bchess 2024-02-23
Left as future work:

* Experiment with different buffer allocation schemes. The improvement with `pinned_use_cuda_host_register` is clearly notable. I'm still exploring how we can achieve similar results by allocating buffers ourselves. And whether anonymous-mmap may help.
* More aggressive thread interrupting behavior in the event of an error
* Potentially copying more than one tensor before synchronizing when they can be read into separate regions of the buffer