@bchess 2024-02-23
Ok @Eta0, PR updated from comments. Most complex change is how _bulk_load yields tensors in order with the supplied keys. For code complexity sake I factored out the caching and ordering behavior. `num_tensors` is also supported again. Take a look!